requirements.txt
Django
djangorestframework
djangorestframework-jwt
requests
factory-boy
pytest-django
